moved some tests into the applicable folder
