+++ src/test/java/edu/wpi/teamR/requestdb/MealRequestTest.java
@@ -1,4 +1,4 @@
-package edu.wpi.teamR.requesttest;
+package edu.wpi.teamR.requestdb;
 
 import java.sql.Connection;
